long or int...that's a question
git-svn-id: http://svn.kcn.cn/repos/kbs/trunk@7945 de4889a5-8ad4-4bf3-bd87-57997c21d515

+++ kbs_bbs/libBBS/article.c
@@ -2,7 +2,7 @@
 #include <utime.h>
 #define DEBUG
 
-int get_effsize_attach(char *ffn, long *att);
+int get_effsize_attach(char *ffn, unsigned int *att);
 int outgo_post(struct fileheader *fh, char *board, char *title, session_t* session)
 {
     FILE *foo;
@@ -1783,7 +1783,7 @@ int get_effsize(char *ffn)
     return abssize;
 }
 
-int get_effsize_attach(char *ffn, long *att)
+int get_effsize_attach(char *ffn, unsigned int *att)
 {
     char *p, *op, *attach;
     long attach_len;
